Experimental cookieless sessions.
We detect if a browser allows us to set a session cookie. If
it doesn't, we rewrite outgoing responses via
DceLti::Middleware::CookielessSessions to include a session key and ID
in forms and URLs.

When a request comes back with a session key and ID in a form parameter
or URL, we use the DceLti::Middleware::CookieShim to "fake" the rails
session out by shimming in a cookie. This allows state to be restored
automatically.

Please see the "Cookieless Sessions" part of the README for more
information and limitations. This behavior is disabled by default.

Includes:

* Set session name and key after a successful launch
* The ability to disable cookieless sessions via config
* Docs.